content/media/webaudio/test/test_oscillatorNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the OscillatorNode interface</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {

  var context = new AudioContext();
  var osc = context.createOscillator();

  is(osc.channelCount, 2, "Oscillator node has 2 input channels by default");
  is(osc.channelCountMode, "max", "Correct channelCountMode for the Oscillator node");
  is(osc.channelInterpretation, "speakers", "Correct channelCountInterpretation for the Oscillator node");
  is(osc.type, "sine", "Correct default type");
  osc.type = "triangle";
  is(osc.type, "triangle", "Can set the type");
  expectException(function() {
    osc.type = "custom";
  }, DOMException.NOT_SUPPORTED_ERR);
  is(osc.type, "triangle", "Cannot set the type to custom");
  is(osc.frequency.value, 440, "Correct default frequency value");
  is(osc.detune.value, 0, "Correct default detine value");

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
